web/requirements.txt
flask
msal
requests
flask-session
python-dotenv
